9180 Move 2 more classes and rename
