+++ test/uk/gov/hmrc/tai/model/domain/CompanyCarBenefitSpec.scala
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.tai.model.domain.formatters
+package uk.gov.hmrc.tai.model.domain
 
-import java.time.LocalDate
 import org.scalatestplus.play.PlaySpec
 import play.api.libs.json.{JsResultException, Json}
 import uk.gov.hmrc.domain.{Generator, Nino}
@@ -24,9 +23,10 @@ import uk.gov.hmrc.tai.model.domain.benefits.CompanyCar.companyCarReadsFromHod
 import uk.gov.hmrc.tai.model.domain.benefits.CompanyCarBenefit.companyCarBenefitReadsFromHod
 import uk.gov.hmrc.tai.model.domain.benefits.{CompanyCar, CompanyCarBenefit}
 
+import java.time.LocalDate
 import scala.util.Random
 
-class CompanyCarBenefitFormattersSpec extends PlaySpec {
+class CompanyCarBenefitSpec extends PlaySpec {
 
   "companyCarBenefitReads" must {
     "be able to read correct companyCarBenefit json" when {
