+++ test/uk/gov/hmrc/tai/model/domain/NpsIabdSummarySpec.scala
@@ -14,17 +14,16 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.tai.model.domain.formatters
+package uk.gov.hmrc.tai.model.domain
 
 import org.scalatestplus.play.PlaySpec
 import play.api.libs.json.{JsArray, JsObject, Json}
 import uk.gov.hmrc.domain.{Generator, Nino}
-import uk.gov.hmrc.tai.model.domain.NpsIabdSummary
 import uk.gov.hmrc.tai.model.domain.NpsIabdSummary.iabdsFromTotalLiabilityReads
 
 import scala.util.Random
 
-class BaseTaxAccountHodFormattersSpec extends PlaySpec {
+class NpsIabdSummarySpec extends PlaySpec {
 
   "iabdsFromTotalLiabilityReads" must {
